refactor: 폴더 구조 정리 및 스타일/타입 유틸 통합
- styles/utils.ts 추가하여 스타일 관련 코드 정리
- 공통 타입 유틸리티를 types/utils.ts로 이동
- 유틸함수를 utils.ts 으로 통합
